# docs/swagger.yaml
definitions:
...
  models.Error:
    properties:
      error:
        type: string
    type: object
  models.FullGear:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      category_top_category_id:
        type: integer
      gear_category_id:
        type: integer
      gear_height:
        type: integer
      gear_id:
        type: integer
      gear_is_container:
        type: boolean
      gear_length:
        type: integer
      gear_manufacture_id:
        type: integer
      gear_name:
        type: string
      gear_size_definition:
        type: string
      gear_status:
        type: boolean
      gear_top_category_id:
        type: integer
      gear_weight:
        type: integer
      gear_width:
        type: integer
      manufacture_id:
        type: integer
      manufacture_name:
        type: string
      top_category_id:
        type: integer
      top_category_name:
        type: string
    type: object
...
  models.ResponsePayload:
    properties:
      current_page:
        type: integer
      item_limit:
        type: integer
      items: {}
      next_page:
        type: string
      prev_page:
        type: string
      total_item_count:
        type: integer
      total_pages:
        type: integer
    type: object
  models.Status:
    properties:
      status:
        type: string
    type: object
...
  models.UserContainerNoID:
    properties:
      user_container_id:
        type: integer
      user_gear_registration_id:
        type: integer
    type: object
...
info:
  contact:
    email: support@sea-shell.no
    name: API Support
  description: This is the API of GoGear
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GoGear API
  version: "1.0"
paths:
...
  /api/v1/container/{container}/delete:
    delete:
      consumes:
      - application/json
      description: Delete user container content with corresponding ID value
      parameters:
      - description: Unique ID of userGear you want to update
        in: path
        name: container
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
        default:
          description: ""
          schema:
            $ref: '#/definitions/models.Error'
      security:
      - BearerAuth: []
      summary: Delete user container content with ID
      tags:
      - User container
  /api/v1/container/{container}/list:
    get:
      consumes:
      - application/json
      description: List all items inside a container registered to a user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 30
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Unique ID of userGear you want to update
        in: path
        name: container
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponsePayload'
            - properties:
                items:
                  items:
                    $ref: '#/definitions/models.FullGear'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: List all items inside a container
      tags:
      - User container
  /api/v1/container/insert:
    put:
      consumes:
      - application/json
      description: Insert user registered gear with corresponding values
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserContainerNoID'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success when all goes well'
          schema:
            $ref: '#/definitions/models.Status'
      security:
      - BearerAuth: []
      summary: Insert user registered gear
      tags:
      - User container
...
securityDefinitions:
  BearerAuth:
    description: Include a server-issued JWT as `Bearer <token>`. Endpoints may require
      either the client or admin audience.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"